requirements.txt
pyyaml
jinja2
colorama
tqdm
rich
